requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Reinforcement Learning & Simulation
gymnasium>=0.28.0
mujoco>=2.3.0
dm-control>=1.0.0

# Utilities
tensorboard>=2.12.0
wandb>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pillow>=9.5.0

# Infrastructure
psutil>=5.9.0
cloudpickle>=2.2.0
